refactor: extract multidropdown props to separate file
