+++ src/Shared/Components/MultiDropdown/index.ts
@@ -1,1 +1,2 @@
 export * from "./MultiDropdown";
+export * from "./propsMultiDropdown";
